+++ src/js/APIs/subway/status.js
@@ -1,0 +1,18 @@
+const STATUS = {
+  EMAIL: {
+    VALID: 200,
+    DUPLICATED: 422,
+  },
+  STATIONS: {
+    VALID: 201,
+    DUPLICATED: 400,
+  },
+  LINES: {
+    DUPLICATED: 400,
+  },
+  SECTIONS: {
+    INVALID: 400,
+  },
+};
+
+export default STATUS;
